# pkg/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CMDB API
  version: 0.1.0
  description: 纯自研 CMDB 核心 API 契约，契约先行，前后端据此并行开发
servers:
  - url: /
    description: 相对路径（与后端同域部署）
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/cis:
    get:
      tags: [cis]
      summary: 查询 CI 列表（支持模型过滤、关键字搜索与分页）
      operationId: listCIs
      parameters:
        - name: model_id
          in: query
          description: 按所属模型过滤
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: 按状态过滤
          required: false
          schema:
            type: string
            enum: [discovered, active, retired]
        - name: keyword
          in: query
          description: 全文关键字（匹配全部属性值，大小写不敏感）
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: CI 分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CIListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [cis]
      summary: 创建 CI
      operationId: createCI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CICreateRequest'
      responses:
        '200':
          description: 创建成功，返回完整 CI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CI'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/cis/{ci_id}:
    get:
      tags: [cis]
      summary: 查询 CI 详情
      operationId: getCI
      parameters:
        - $ref: '#/components/parameters/CIId'
      responses:
        '200':
          description: CI 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CI'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [cis]
      summary: 更新 CI（部分字段）
      operationId: patchCI
      parameters:
        - $ref: '#/components/parameters/CIId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CIPatchRequest'
      responses:
        '200':
          description: 更新成功，返回完整 CI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CI'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/cis/{ci_id}/relations:
    get:
      tags: [cis]
      summary: 查询 CI 的关系列表
      operationId: listCIRelations
      parameters:
        - $ref: '#/components/parameters/CIId'
      responses:
        '200':
          description: 关系列表（含对端 CI 摘要）
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CIRelation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [cis]
      summary: 创建 CI 关系
      operationId: createCIRelation
      description: 在当前 CI 与对端 CI 间建立关系；关系编码须在所属模型的关系定义内，对端 CI 模型须匹配 target_model。cardinality 为 one_to_one 时自动替换同编码同方向的既有关系（用于改挂，如机柜换机房）
      parameters:
        - $ref: '#/components/parameters/CIId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CIRelationCreateRequest'
      responses:
        '200':
          description: 创建成功，返回关系（含对端 CI 摘要）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CIRelation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/dbms/eol-report:
    get:
      tags: [dbms]
      summary: 数据库实例 EOL 清单导出
      description: |
        扫描 db_instance CI（排除已退役），按组件类型与版本前缀过滤，
        输出实例地址/版本/角色/所属集群/所属应用（沿 depends_on 入向反查 biz_app）/负责人。
        format=csv 时返回带 BOM 的 CSV（表头与 JSON 字段同名）。
        数据范围（F-005）：受约束用户仅导出归属闭包内的实例。
      operationId: getDBMSEOLReport
      parameters:
        - name: component
          in: query
          description: 组件类型精确过滤（如 mysql/redis）
          required: false
          schema:
            type: string
        - name: version_prefix
          in: query
          description: 版本前缀过滤（如 5.7）
          required: false
          schema:
            type: string
        - name: format
          in: query
          description: 输出格式，默认 json
          required: false
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: EOL 清单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EOLReportResponse'
            text/csv:
              schema:
                type: string
                description: 带 BOM 的 CSV，表头 instance_addr,component_type,version,role,cluster_name,app_name,app_owner
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: meridian_token
      description: 登录成功后由服务端写入的 httpOnly JWT 会话 cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 采集器等非浏览器客户端使用的 Bearer JWT(登录接口响应体中的 token)
  parameters:
...
    CIId:
      name: ci_id
      in: path
      description: CI ID
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      description: 页码，从 1 开始
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      description: 每页条数
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  responses:
    BadRequest:
      description: 请求参数或校验规则不通过
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未登录或会话已失效(错误码 UNAUTHORIZED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 已登录但缺少所需权限点(错误码 FORBIDDEN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 机器可读错误码
          example: VALIDATION_FAILED
        message:
          type: string
          description: 人类可读错误信息
          example: 属性 hostname 为必填项
        details:
          type: object
          additionalProperties: true
          description: 逐字段校验错误等补充信息
...
    CI:
      type: object
      required: [id, model_id, attributes, status, source, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          description: CI ID
        model_id:
          type: string
          format: uuid
          description: 所属模型 ID
        attributes:
          type: object
          additionalProperties: true
          description: 属性键值对，键为属性编码，取值须通过模型校验规则
        status:
          type: string
          enum: [discovered, active, retired]
          description: 生命周期状态
        source:
          type: string
          description: 建档来源（manual/采集器标识）
          example: n9e
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CICreateRequest:
      type: object
      required: [model_id, attributes]
      properties:
        model_id:
          type: string
          format: uuid
        attributes:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [discovered, active, retired]
          default: active
        source:
          type: string
          default: manual
    CIPatchRequest:
      type: object
      description: 全字段可选，attributes 为增量合并
      properties:
        attributes:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [discovered, active, retired]
        source:
          type: string
    CIListResponse:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CI'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    CIRelation:
      type: object
      required: [relation_code, direction, peer_ci]
      properties:
        relation_code:
          type: string
          description: 关系编码
          example: runs_on
        direction:
          type: string
          enum: [outgoing, incoming]
          description: 相对当前 CI 的方向
        peer_ci:
          $ref: '#/components/schemas/CI'
    CIRelationCreateRequest:
      type: object
      required: [relation_code, peer_ci_id]
      properties:
        relation_code:
          type: string
          description: 关系编码（须在所属模型的关系定义内）
          example: located_in
        peer_ci_id:
          type: string
          format: uuid
          description: 对端 CI ID（其模型须匹配关系定义的 target_model）
...
    EOLReportItem:
      type: object
      required: [instance_addr, component_type, version, role, cluster_name, app_name, app_owner]
      properties:
        instance_addr:
          type: string
          description: 实例地址（ip:port）
        component_type:
          type: string
          description: 组件类型（mysql/redis 等）
        version:
          type: string
        role:
          type: string
          description: 实例角色（master/slave/standalone）
        cluster_name:
          type: string
          description: 所属集群
        app_name:
          type: string
          description: 所属应用名称（沿 depends_on 入向反查 biz_app，无归属为空串）
        app_owner:
          type: string
          description: 应用负责人
    EOLReportResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EOLReportItem'
...